Split next and common context
